+++ NEWS.md
@@ -1,5 +1,13 @@
 # learndown 1.0.4
 
+- The fields score, max and grade are added in all events tables with
+grade = score/max. The solution for Shiny applications can now use a range for
+numeric values and a series of correct solutions for text. Also a `max_score =`
+argument added to `trackSubmit()` allows to redefine the maximum score for a
+Shiny application.
+
+# learndown 1.0.4
+
 - Functions for learndown learnr applications added.
 - `config()` function added.
 - Reworked database scheme, including course and institution